feat(logging): add MdcCleanupFilter for improved logging context management
